Move TestPubSub
